+++ Changelog.md
@@ -10,6 +10,9 @@
 * Array keys (and shapes in general) are now tracked more granularly when backpropagating the effects of a function call.
 * (MW) Analyze the `$rows` argument to `Database::insert()` more accurately.
 
+### Bug fixes
+* Fixed a bug where `*-taint` annotations in an interface method were only inherited by the method implementation in children classes.
+
 ## v5.0.0
 ### Breaking changes
 * The raw_param taint flag was removed; error reporting is now sufficiently good that this is no longer needed, and can be treated as normal exec.